ML/requirements.txt
tensorflow-cpu==2.17.1
numpy>=1.24.3,<2.0.0
pandas>=2.1.4
scikit-learn>=1.3.2
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6